--- client/src/components/Player.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-class Player extends React.Component {
-  render(props) {
-    return (
-      <div className="player">
-        <h3>{this.props.name}</h3>
-        <h4>{this.props.country}</h4>
-      </div>
-    );
-  }
-}
-export default Player;
