+++ client/src/index.css
@@ -60,7 +60,7 @@ code {
 
 .navbar {
   align-items: center;
-  border-bottom: 1px solid rgb(221, 221, 221);
+  border-bottom: 1px solid rgb(143, 18, 18);
   display: flex;
   height: 70px;
   justify-content: space-between;
